fix(input-box): refactor namming
